+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Support for new environments `preview` and `preprod`.
+- Supported derivation of root private key using second factor mnemonic or its base16/base64/plain text equivalent
 
 ## [3.12.0] - 2022-08-17